Send email when Google account is connected

+++ CloudDataProtection/CloudDataProtection.Gateway/Data/IAuthenticationRepository.cs
@@ -5,7 +5,7 @@ namespace CloudDataProtection.Data
 {
     public interface IAuthenticationRepository
     {
-        Task<User> Get(int id);
+        Task<User> Get(long id);
         Task<User> Get(string email);
 
         Task Create(User user);
